+++ scripts/version_bump.sh
@@ -10,7 +10,7 @@ current_state=$(git rev-parse HEAD)
 changes=$(git diff --name-only $last_release $current_state)
 
 # Check if any public API changes have been made
-api_changes=$(echo "$changes" | grep -v "internal" | grep ".go")
+file_changes=$(echo "$changes" | grep -v "internal" | grep ".go")
 
 # Get the current version
 version=$(echo $last_release | sed 's/v//')
@@ -26,19 +26,14 @@ else
   bug_fix=false
 fi
 
-# Check if the last version contains "-preview"
-if [[ $last_release == *"-preview"* ]]; then
-  # If the minor version is 0, increment the major version
-  if [[ $minor_version == 0 ]]; then
-    major_version=$((major_version + 1))
-  # If the patch version is 0, increment the minor version
-  elif [[ $patch_version == 0 ]]; then
-    minor_version=$((minor_version + 1))
-  fi
-  # Reset the patch version to 0
-  patch_version=0
+# See if any signature changes happen
+last_revision=$(git rev-parse $last_release)
+for file in $file_changes; do
+  echo $(git diff "$last_revision..$current_state" -- "$file") >> /dev/null
+done
+
 # If any public API changes have been made, increment the minor version
-elif [[ -n $api_changes ]]; then
+if [[ -n $file_changes ]]; then
   minor_version=$((minor_version + 1))
   patch_version=0
 # If the current branch is a bug-fix and the minor version isn't incremented, increment the patch version