+++ changelog.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- `core.PageIterator2[T].Previous` ([Issue #60](https://github.com/michaeldcanady/servicenow-sdk-go/issues/60))
+- `core.PageIterator2[T].First` ([Issue #60](https://github.com/michaeldcanady/servicenow-sdk-go/issues/60))
+
 ### Changed
 
 ### Deprecated